+++ src/pages/HomePage/HomePage.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@/components/ui/button';
 
-function Home() {
+function HomePage() {
   return (
     <>
       <div>
@@ -16,4 +16,4 @@ function Home() {
   );
 }
 
-export default Home;
+export default HomePage;
