+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,5 +1,5 @@
-function NotFound() {
+function NotFoundPage() {
   return <div>페이지를 찾을 수 없습니다.</div>;
 }
 
-export default NotFound;
+export default NotFoundPage;
