+++ src/router.tsx
@@ -1,10 +1,10 @@
 import { createBrowserRouter } from 'react-router';
-import AppRoot from './AppRoot';
-import NotFound from './pages/NotFound/NotFound';
-import Companies from './pages/Companies/Companies';
-import Login from './pages/Login';
-import Signup from './pages/Signup';
-import Home from './pages/Home/Home';
+import AppRoot from './components/layout/AppRoot.tsx/AppRoot';
+import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
+import CompaniesPage from './pages/CompaniesPage/CompaniesPage';
+import LoginPage from './pages/LoginPage/LoginPage';
+import SignupPage from './pages/SignupPage/SignupPage';
+import HomePage from './pages/HomePage/HomePage';
 import MainRoot from './components/layout/MainRoot/MainRoot';
 
 /**
@@ -14,17 +14,17 @@ const router = createBrowserRouter([
   {
     path: '/',
     element: <AppRoot />,
-    errorElement: <NotFound />,
+    errorElement: <NotFoundPage />,
     children: [
       {
         element: <MainRoot />,
         children: [
-          { index: true, element: <Home /> },
-          { path: 'companies', element: <Companies /> },
+          { index: true, element: <HomePage /> },
+          { path: 'companies', element: <CompaniesPage /> },
         ],
       },
-      { path: 'login', element: <Login /> },
-      { path: 'signup', element: <Signup /> },
+      { path: 'login', element: <LoginPage /> },
+      { path: 'signup', element: <SignupPage /> },
     ],
   },
 ]);
